backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
langchain==0.2.0
langchain-groq==0.1.3
langchain-community==0.2.0
langchain-huggingface==0.0.3
chromadb==0.5.0
sentence-transformers==3.0.0
PyPDF2==3.0.1
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
twilio==9.1.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
requests==2.32.2
python-multipart==0.0.9
